+++ bin.js
@@ -2,6 +2,7 @@
 
 import sade from 'sade'
 import open from 'open'
+import updateNotifier from 'update-notifier'
 import { getPkg } from './lib.js'
 import {
   accessClaim,
@@ -31,10 +32,14 @@ import {
   uploadRemove
 } from './can.js'
 
+const pkg = getPkg()
+
+updateNotifier({ pkg }).notify()
+
 const cli = sade('w3')
 
 cli
-  .version(getPkg().version)
+  .version(pkg.version)
   .example('authorize user@example.com')
   .example('up path/to/files')
 
@@ -51,6 +56,8 @@ cli.command('up <file>')
   .option('--no-wrap', 'Don\'t wrap input files with a directory.', false)
   .option('-H, --hidden', 'Include paths that start with ".".')
   .option('-c, --car', 'File is a CAR file.', false)
+  .option('--json', 'Format as newline delimited JSON')
+  .option('--verbose', 'Output more details.')
   .option('--shard-size', 'Shard uploads into CAR files of approximately this size in bytes.')
   .option('--concurrent-requests', 'Send up to this many CAR shards concurrently.')
   .action(upload)